scholarship_chatbot_project/scholarship_chatbot/backend/requirements.txt
fastapi
uvicorn
pandas
faiss-cpu
sentence-transformers
